Theoretically Fixed the API
Dan Scribner provided code for requesting a single entry from the API, so I modified the code to use only that one entry. Trouble is, I can't figure out how to test it. Given the simplicity of the changes and the fact that the old code still exists in the comments, I'm committing it anyway.